move XML resource parsers to res/parser
